refactor xlsx dynamodb fixture
